requirements.txt
# ── Backend ────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
websockets>=14.0
aiofiles>=24.0

# ── Transcription ──────────────────────────
vosk>=0.3.45
faster-whisper>=1.1.0

# ── Video/Image processing ─────────────────
ffmpeg-python>=0.2.0
numpy>=1.24.0
Pillow>=10.0.0

# ── Desktop app ────────────────────────────
pystray>=0.19.5
customtkinter>=5.2.0

# ── Misc ───────────────────────────────────
requests>=2.31.0
